More rearranging.
