+++ pb-coroutines-retry/src/main/kotlin/io/provenance/kafka/coroutines/retry/ByteArrayExtensions.kt
@@ -7,4 +7,3 @@ fun Int.toByteArray() =
 
 fun ByteArray.toInt() =
     ByteBuffer.wrap(this).int
-
